Address inconsistencies between MixedBehaviorProfile and MixedStrategyProfile.
This cleans up a few inconsistencies in the interfaces to the two mixed profile classes:

* Implementation details are now hidden (MixedBehaviorProfile is not derived from DVector)
* MixedProfileLength/BehaviorProfileLength are now used instead of Length
* Non-const casting to vectors is removed (breaks cacheing mechanisms)
* Remove unused operations

+++ src/solvers/gnm/gnm.cc
@@ -34,7 +34,7 @@ NashGNMStrategySolver::ToProfile(const Game &p_game,
 				 const cvector &p_pert)
 {
   MixedStrategyProfile<double> msp = p_game->NewMixedStrategyProfile(0.0);
-  for (int i = 1; i <= msp.MixedProfileLength(); i++) {
+  for (size_t i = 1; i <= msp.MixedProfileLength(); i++) {
     msp[i] = p_pert[i-1];
   }
   return msp;
